+++ Assets/Scripts/MVC/Models/Songs/SongSettings.cs
@@ -3,7 +3,7 @@
 public class SongSettings : ISongSettings
 {
     public string Id { get; set; }
-    public double Bpm { get; set; }
+    public float Bpm { get; set; }
     public float ApproachRate { get; set; }
     public float Difficulty { get; set; }
     public float StartingTime { get; set; }
